+++ tests/test_eof.py
@@ -32,6 +32,7 @@ def _get_header(obj):
         return obj.earth_explorer_header
 
 
+# === High level functions ====================================================
 @pytest.mark.parametrize("filename", AUX_ORBRES)
 def test_load_from_filename(filename):
     obj = xseof.load(str(filename))
@@ -87,11 +88,26 @@ def test_from_bytes(filename):
     assert header.fixed_header.file_name == filename.stem
 
 
+# === aux_orbdop functions ====================================================
 @pytest.mark.parametrize("filename", AUX_ORBDOP)
 def test_load_aux_orbdop(filename):
     obj = xseof.aux_orbdop.load(filename)
     header = _get_header(obj)
     assert header.fixed_header.file_name == filename.stem
+
+
+@pytest.mark.parametrize("filename", AUX_ORBDOP[:1])
+def test_load_fd_aux_orbdop(filename):
+    with open(filename) as fd:
+        obj = xseof.aux_orbdop.load(fd)
+    header = _get_header(obj)
+    assert header.fixed_header.file_name == filename.stem
+
+
+@pytest.mark.parametrize("filename", INT_ATTREF[:1])
+def test_load_invalid_aux_orbdop(filename):
+    with pytest.raises(xseof.ParserError):
+        xseof.aux_orbdop.load(filename)
 
 
 @pytest.mark.parametrize("filename", AUX_ORBDOP)
@@ -110,11 +126,26 @@ def test_from_bytes_aux_orbdop(filename):
     assert header.fixed_header.file_name == filename.stem
 
 
+# === aux_orbdor functions ====================================================
 @pytest.mark.parametrize("filename", AUX_ORBDOR)
 def test_load_aux_orbdor(filename):
     obj = xseof.aux_orbdor.load(filename)
     header = _get_header(obj)
     assert header.fixed_header.file_name == filename.stem
+
+
+@pytest.mark.parametrize("filename", AUX_ORBDOR[:1])
+def test_load_fd_aux_orbdor(filename):
+    with open(filename) as fd:
+        obj = xseof.aux_orbdor.load(fd)
+    header = _get_header(obj)
+    assert header.fixed_header.file_name == filename.stem
+
+
+@pytest.mark.parametrize("filename", INT_ATTREF[:1])
+def test_load_invalid_aux_orbdor(filename):
+    with pytest.raises(xseof.ParserError):
+        xseof.aux_orbdor.load(filename)
 
 
 @pytest.mark.parametrize("filename", AUX_ORBDOR)
@@ -133,11 +164,26 @@ def test_from_bytes_aux_orbdor(filename):
     assert header.fixed_header.file_name == filename.stem
 
 
+# === aux_orbres functions ====================================================
 @pytest.mark.parametrize("filename", AUX_ORBRES)
 def test_load_aux_orbres(filename):
     obj = xseof.aux_orbres.load(filename)
     header = _get_header(obj)
     assert header.fixed_header.file_name == filename.stem
+
+
+@pytest.mark.parametrize("filename", AUX_ORBRES[:1])
+def test_load_fd_aux_orbres(filename):
+    with open(filename) as fd:
+        obj = xseof.aux_orbres.load(fd)
+    header = _get_header(obj)
+    assert header.fixed_header.file_name == filename.stem
+
+
+@pytest.mark.parametrize("filename", INT_ATTREF[:1])
+def test_load_invalid_aux_orbres(filename):
+    with pytest.raises(xseof.ParserError):
+        xseof.aux_orbres.load(filename)
 
 
 @pytest.mark.parametrize("filename", AUX_ORBRES)
@@ -156,11 +202,26 @@ def test_from_bytes_aux_orbres(filename):
     assert header.fixed_header.file_name == filename.stem
 
 
+# === int_attref functions ====================================================
 @pytest.mark.parametrize("filename", INT_ATTREF)
 def test_load_int_attref(filename):
     obj = xseof.int_attref.load(filename)
     header = _get_header(obj)
     assert header.fixed_header.file_name == filename.stem
+
+
+@pytest.mark.parametrize("filename", INT_ATTREF[:1])
+def test_load_fd_int_attref(filename):
+    with open(filename) as fd:
+        obj = xseof.int_attref.load(fd)
+    header = _get_header(obj)
+    assert header.fixed_header.file_name == filename.stem
+
+
+@pytest.mark.parametrize("filename", AUX_ORBDOP[:1])
+def test_load_invalid_int_attref(filename):
+    with pytest.raises(xseof.ParserError):
+        xseof.int_attref.load(filename)
 
 
 @pytest.mark.parametrize("filename", INT_ATTREF)
@@ -179,11 +240,26 @@ def test_from_bytes_int_attref(filename):
     assert header.fixed_header.file_name == filename.stem
 
 
+# === mpl_orbpre functions ====================================================
 @pytest.mark.parametrize("filename", MPL_ORBPRE)
 def test_load_mpl_orbpre(filename):
     obj = xseof.mpl_orbpre.load(filename)
     header = _get_header(obj)
     assert header.fixed_header.file_name == filename.stem
+
+
+@pytest.mark.parametrize("filename", MPL_ORBPRE[:1])
+def test_load_fd_mpl_orbpre(filename):
+    with open(filename) as fd:
+        obj = xseof.mpl_orbpre.load(fd)
+    header = _get_header(obj)
+    assert header.fixed_header.file_name == filename.stem
+
+
+@pytest.mark.parametrize("filename", INT_ATTREF[:1])
+def test_load_invalid_mpl_orbpre(filename):
+    with pytest.raises(xseof.ParserError):
+        xseof.mpl_orbpre.load(filename)
 
 
 @pytest.mark.parametrize("filename", MPL_ORBPRE)
@@ -202,11 +278,26 @@ def test_from_bytes_mpl_orbpre(filename):
     assert header.fixed_header.file_name == filename.stem
 
 
+# === mpl_orbref functions ====================================================
 @pytest.mark.parametrize("filename", MPL_ORBREF)
 def test_load_mpl_orbref(filename):
     obj = xseof.mpl_orbref.load(filename)
     header = _get_header(obj)
     assert header.fixed_header.file_name == filename.stem
+
+
+@pytest.mark.parametrize("filename", MPL_ORBREF[:1])
+def test_load_fd_mpl_orbref(filename):
+    with open(filename) as fd:
+        obj = xseof.mpl_orbref.load(fd)
+    header = _get_header(obj)
+    assert header.fixed_header.file_name == filename.stem
+
+
+@pytest.mark.parametrize("filename", INT_ATTREF[:1])
+def test_load_invalid_mpl_orbref(filename):
+    with pytest.raises(xseof.ParserError):
+        xseof.mpl_orbref.load(filename)
 
 
 @pytest.mark.parametrize("filename", MPL_ORBREF)
@@ -225,6 +316,23 @@ def test_from_bytes_mpl_orbref(filename):
     assert header.fixed_header.file_name == filename.stem
 
 
+# === strict parameter ========================================================
+@pytest.mark.parametrize("module", [
+    xseof,
+    xseof.aux_orbdop,
+    xseof.aux_orbdor,
+    xseof.aux_orbres,
+    xseof.int_attref,
+    xseof.mpl_orbpre,
+    xseof.mpl_orbref,
+])
+def test_from_invalid_string(module):
+    data = "dummy"
+    with pytest.raises(xseof.ParserError):
+        module.from_string(data)
+
+
+# === no-strict parsing =======================================================
 @pytest.mark.parametrize("filename", S1X_XXXORB)
 def test_load_s1_orb(filename):
     pytest.importorskip("lxml")
@@ -272,3 +380,10 @@ def test_load_s1_orb_from_bytes_strict(filename):
     xml_bytes = filename.read_bytes()
     with pytest.raises(xseof.ParserError):
         xseof.from_string(xml_bytes, strict=True)
+
+
+def test_load_s1_orb_from_invalid_bytes():
+    pytest.importorskip("lxml")
+    xml_bytes = "<dummy><dummytag>aaa</dummytag></dummy>"
+    with pytest.raises(xseof.ParserError):
+        xseof.from_string(xml_bytes)