+++ xseof/aux_orbdop/__init__.py
@@ -31,6 +31,9 @@ from .eo_oper_aux_orbdop_0201 import (  # noqa: F401
 from .eo_oper_aux_orbdop_0202 import (  # noqa: F401
     EarthExplorerFile as AuxOrbDopFileV0202,
 )
+from .eo_oper_aux_orbdop_0203 import (  # noqa: F401
+    EarthExplorerFile as AuxOrbDopFileV0203,
+)
 from .eo_oper_aux_orbdop_0300 import (  # noqa: F401
     EarthObservationFile as AuxOrbDopFileV0300,
 )