+++ xseof/mpl_orbpre/__init__.py
@@ -31,6 +31,9 @@ from .eo_oper_mpl_orbpre_0201 import (  # noqa: F401
 from .eo_oper_mpl_orbpre_0202 import (  # noqa: F401
     EarthExplorerFile as MplOrbPreFileV0202,
 )
+from .eo_oper_mpl_orbpre_0203 import (  # noqa: F401
+    EarthExplorerFile as MplOrbPreFileV0203,
+)
 from .eo_oper_mpl_orbpre_0300 import (  # noqa: F401
     EarthObservationFile as MplOrbPreFileV0300,
 )